Update WebApplication template for beta7. Closes #317
- update project structure
- add new content (Models/ViewModels/etc)
- rename content
- remove files
- add new tests
- update existing tests
- update Razor templates
- update dnx usage hints
