+++ templates/projects/web/Startup.cs
@@ -7,19 +7,14 @@ using Microsoft.AspNet.Authentication.Google;
 using Microsoft.AspNet.Authentication.MicrosoftAccount;
 using Microsoft.AspNet.Authentication.Twitter;
 using Microsoft.AspNet.Builder;
-using Microsoft.AspNet.Diagnostics;
 using Microsoft.AspNet.Diagnostics.Entity;
 using Microsoft.AspNet.Hosting;
-using Microsoft.AspNet.Http;
-using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.EntityFramework;
-using Microsoft.AspNet.Routing;
 using Microsoft.Data.Entity;
+using Microsoft.Dnx.Runtime;
 using Microsoft.Framework.Configuration;
 using Microsoft.Framework.DependencyInjection;
 using Microsoft.Framework.Logging;
-using Microsoft.Framework.Logging.Console;
-using Microsoft.Framework.Runtime;
 using <%= namespace %>.Models;
 using <%= namespace %>.Services;
 
@@ -43,25 +38,19 @@ namespace <%= namespace %>
             }
             builder.AddEnvironmentVariables();
             Configuration = builder.Build();
+            Configuration["Data:DefaultConnection:ConnectionString"] = $@"Data Source={appEnv.ApplicationBasePath}\<%= namespace %>.db";
         }
 
-        public IConfiguration Configuration { get; set; }
+        public IConfigurationRoot Configuration { get; set; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
             // Add Entity Framework services to the services container.
-            // In this release there is no data provider for non Windows environment so this app
-            // uses the InMemory Store. This is coming in a future release of the Framework.
             services.AddEntityFramework()
-                .AddInMemoryDatabase()
-                .AddDbContext<ApplicationDbContext>(options => options.UseInMemoryDatabase());
-
-            // Use the following code to register EntityFramework services for SqlServer to the container.
-            //services.AddEntityFramework()
-            //   .AddSqlServer()
-            //   .AddDbContext<ApplicationDbContext>(options =>
-            //       options.UseSqlServer(Configuration["Data:DefaultConnection:ConnectionString"]));
+                .AddSqlite()
+                .AddDbContext<ApplicationDbContext>(options =>
+                    options.UseSqlite(Configuration["Data:DefaultConnection:ConnectionString"]));
 
             // Add Identity services to the services container.
             services.AddIdentity<ApplicationUser, IdentityRole>()
@@ -100,6 +89,7 @@ namespace <%= namespace %>
         {
             loggerFactory.MinimumLevel = LogLevel.Information;
             loggerFactory.AddConsole();
+            loggerFactory.AddDebug();
 
             // Configure the HTTP request pipeline.
