+++ templates/projects/web/ViewModels/Manage/FactorViewModel.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace <%= namespace %>.ViewModels.Manage
+{
+    public class FactorViewModel
+    {
+        public string Purpose { get; set; }
+    }
+}
