+++ templates/projects/web/Views/Account/ExternalLoginConfirmation.cshtml
@@ -1,4 +1,4 @@
-﻿@model ExternalLoginConfirmationViewModel
+@model ExternalLoginConfirmationViewModel
 @{
     ViewData["Title"] = "Register";
 }
@@ -6,7 +6,7 @@
 <h2>@ViewData["Title"].</h2>
 <h3>Associate your @ViewData["LoginProvider"] account.</h3>
 
-<form asp-controller="Account" asp-action="ExternalLoginConfirmation" asp-route-returnurl="@ViewData["ReturnUrl"] method="post" class="form-horizontal" role="form">
+<form asp-controller="Account" asp-action="ExternalLoginConfirmation" asp-route-returnurl="@ViewData["ReturnUrl"]" method="post" class="form-horizontal" role="form">
     <h4>Association Form</h4>
     <hr />
     <div asp-validation-summary="ValidationSummary.All" class="text-danger"></div>
