+++ templates/projects/web/Views/Manage/RemoveLogin.cshtml
@@ -1,4 +1,4 @@
-﻿@model ICollection<Microsoft.AspNet.Identity.UserLoginInfo>
+@model ICollection<Microsoft.AspNet.Identity.UserLoginInfo>
 @{
     ViewData["Title"] = "Remove Login";
 }
@@ -15,7 +15,7 @@
                     <td>
                         @if ((bool)ViewData["ShowRemoveButton"])
                         {
-                            <form asp-controller="Manage" asp-action="RemoveLogin" asp-route-returnurl="@ViewData["ReturnUrl"] method="post" class="form-horizontal" role="form">
+                            <form asp-controller="Manage" asp-action="RemoveLogin" asp-route-returnurl="@ViewData["ReturnUrl"]" method="post" class="form-horizontal" role="form">
                                 <div>
                                     <input asp-for="@account.LoginProvider" type="hidden" />
                                     <input asp-for="@account.ProviderKey" type="hidden" />
