+++ templates/projects/web/Views/Manage/SetPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model SetPasswordViewModel
+@model SetPasswordViewModel
 @{
     ViewData["Title"] = "Set Password";
 }
@@ -8,7 +8,7 @@
     account so you can log in without an external login.
 </p>
 
-<form asp-controller="Manage" asp-action="SetPassword" asp-route-returnurl="@ViewData["ReturnUrl"] method="post" class="form-horizontal" role="form">
+<form asp-controller="Manage" asp-action="SetPassword" asp-route-returnurl="@ViewData["ReturnUrl"]" method="post" class="form-horizontal" role="form">
     <h4>Set your password</h4>
     <hr />
     <div asp-validation-summary="ValidationSummary.All" class="text-danger"></div>
