+++ templates/projects/web/Views/Manage/VerifyPhoneNumber.cshtml
@@ -1,11 +1,11 @@
-﻿@model VerifyPhoneNumberViewModel
+@model VerifyPhoneNumberViewModel
 @{
     ViewData["Title"] = "Verify Phone Number";
 }
 
 <h2>@ViewData["Title"].</h2>
 
-<form asp-controller="Manage" asp-action="VerifyPhoneNumber" asp-route-returnurl="@ViewData["ReturnUrl"] method="post" class="form-horizontal" role="form">
+<form asp-controller="Manage" asp-action="VerifyPhoneNumber" asp-route-returnurl="@ViewData["ReturnUrl"]" method="post" class="form-horizontal" role="form">
     <input asp-for="PhoneNumber" type="hidden" />
     <h4>Add a phone number.</h4>
     <h5>@ViewData["Status"]</h5>
