+++ templates/projects/web/Views/Shared/_ValidationScriptsPartial.cshtml
@@ -1,9 +1,9 @@
-﻿<environment names="Development">
+<environment names="Development">
     <script src="~/lib/jquery-validation/jquery.validate.js"></script>
     <script src="~/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.js"></script>
 </environment>
 <environment names="Staging,Production">
-    <script src="//ajax.aspnetcdn.com/ajax/jquery.validation/1.11.1/jquery.validate.min.js"
+    <script src="//ajax.aspnetcdn.com/ajax/jquery.validate/1.11.1/jquery.validate.min.js"
             asp-fallback-src="~/lib/jquery-validation/jquery.validate.js"
             asp-fallback-test="window.jquery && window.jquery.validator">
     </script>
