+++ templates/projects/web/Views/_ViewImports.cshtml
@@ -1,4 +1,6 @@
-﻿@using <%= namespace %>
-@using <%= namespace %>.Models
+@using WebApplication
+@using WebApplication.Models
+@using WebApplication.ViewModels.Account
+@using WebApplication.ViewModels.Manage
 @using Microsoft.AspNet.Identity
 @addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
