+++ templates/projects/web/config.json
@@ -1,10 +1,2 @@
 {
-  "AppSettings": {
-    "SiteTitle": "<%= namespace %>"
-  },
-  "Data": {
-    "DefaultConnection": {
-      "ConnectionString": "Server=(localdb)\\mssqllocaldb;Database=aspnet5-<%= namespace %>-<%= guid %>;Trusted_Connection=True;MultipleActiveResultSets=true"
-    }
-  }
 }
