+++ templates/projects/web/project.json
@@ -4,34 +4,35 @@
   "version": "1.0.0-*",
 
   "dependencies": {
-    "EntityFramework.InMemory": "7.0.0-beta6",
-    "EntityFramework.Commands": "7.0.0-beta6",
-    "Microsoft.AspNet.Mvc": "6.0.0-beta6",
-    "Microsoft.AspNet.Mvc.TagHelpers": "6.0.0-beta6",
-    "Microsoft.AspNet.Authentication.Cookies": "1.0.0-beta6",
-    "Microsoft.AspNet.Authentication.Facebook": "1.0.0-beta6",
-    "Microsoft.AspNet.Authentication.Google": "1.0.0-beta6",
-    "Microsoft.AspNet.Authentication.MicrosoftAccount": "1.0.0-beta6",
-    "Microsoft.AspNet.Authentication.Twitter": "1.0.0-beta6",
-    "Microsoft.AspNet.Diagnostics": "1.0.0-beta6",
-    "Microsoft.AspNet.Diagnostics.Entity": "7.0.0-beta6",
-    "Microsoft.AspNet.Identity.EntityFramework": "3.0.0-beta6",
-    "Microsoft.AspNet.Server.IIS": "1.0.0-beta6",
-    "Microsoft.AspNet.Server.WebListener": "1.0.0-beta6",
-    "Microsoft.AspNet.StaticFiles": "1.0.0-beta6",
-    "Microsoft.AspNet.Tooling.Razor": "1.0.0-beta6",
-    "Microsoft.Framework.Configuration.Abstractions": "1.0.0-beta6",
-    "Microsoft.Framework.Configuration.Json": "1.0.0-beta6",
-    "Microsoft.Framework.Configuration.UserSecrets": "1.0.0-beta6",
-    "Microsoft.Framework.Logging": "1.0.0-beta6",
-    "Microsoft.Framework.Logging.Console": "1.0.0-beta6",
-    "Microsoft.AspNet.Server.Kestrel": "1.0.0-beta6",
-    "Kestrel": "1.0.0-beta6"
+    "EntityFramework.SQLite": "7.0.0-beta7",
+    "EntityFramework.Commands": "7.0.0-beta7",
+    "Microsoft.AspNet.Mvc": "6.0.0-beta7",
+    "Microsoft.AspNet.Mvc.TagHelpers": "6.0.0-beta7",
+    "Microsoft.AspNet.Authentication.Cookies": "1.0.0-beta7",
+    "Microsoft.AspNet.Authentication.Facebook": "1.0.0-beta7",
+    "Microsoft.AspNet.Authentication.Google": "1.0.0-beta7",
+    "Microsoft.AspNet.Authentication.MicrosoftAccount": "1.0.0-beta7",
+    "Microsoft.AspNet.Authentication.Twitter": "1.0.0-beta7",
+    "Microsoft.AspNet.Diagnostics": "1.0.0-beta7",
+    "Microsoft.AspNet.Diagnostics.Entity": "7.0.0-beta7",
+    "Microsoft.AspNet.Identity.EntityFramework": "3.0.0-beta7",
+    "Microsoft.AspNet.Server.IIS": "1.0.0-beta7",
+    "Microsoft.AspNet.Server.WebListener": "1.0.0-beta7",
+    "Microsoft.AspNet.StaticFiles": "1.0.0-beta7",
+    "Microsoft.AspNet.Tooling.Razor": "1.0.0-beta7",
+    "Microsoft.Framework.Configuration.Abstractions": "1.0.0-beta7",
+    "Microsoft.Framework.Configuration.Json": "1.0.0-beta7",
+    "Microsoft.Framework.Configuration.UserSecrets": "1.0.0-beta7",
+    "Microsoft.Framework.Logging": "1.0.0-beta7",
+    "Microsoft.Framework.Logging.Console": "1.0.0-beta7",
+    "Microsoft.Framework.Logging.Debug" : "1.0.0-beta7",
+    "Microsoft.AspNet.Server.Kestrel": "1.0.0-beta7"
   },
 
   "commands": {
-    "kestrel": "Microsoft.AspNet.Hosting --server Kestrel --config hosting.ini",
-    "web": "Microsoft.AspNet.Hosting --server Microsoft.AspNet.Server.WebListener --config hosting.ini"
+    "kestrel": "Microsoft.AspNet.Hosting --server Microsoft.AspNet.Server.Kestrel --config hosting.ini",
+    "web": "Microsoft.AspNet.Hosting --server Microsoft.AspNet.Server.WebListener --config hosting.ini",
+    "ef": "EntityFramework.Commands"
   },
 
   "frameworks": {
